streaming/client/requirements.txt
aiortc>=1.6.0
aiohttp>=3.9.0
python-socketio[asyncio]>=5.10.0
av>=11.0.0
picamera2>=0.3.17
numpy>=1.24.0
opencv-python-headless>=4.8.0
pyyaml>=6.0.0
